+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 Following Semantic Versioning 2.
 
+## Version 0.0.2
+- Make module compatible with Decidim v0.30.
+- Add i18n-tasks gem and specs.
+- Bump Ruby to 3.3.4.
+
 ## Version 0.0.1
 - Initialize module